Add files to prepare for siamese NN
All of the `.py` files except for `kat.py` are for training
the siamese neural network that will be used for
image similarity computation down the line. For now, the
current state of `kat.py` is sufficient for obtaining
mostly reliable timestamps. Next commit will add
command line argument for the video to make it easier.