Complete siamese NN version
`kat_nn.py` contains the auto-timestamper that utilises a siamese neural
network to make the character portrait comparison. At the moment, it
doesn't perform very well compared to using Stuctural Similarity. This
is not surprising, as I would need more data to improve the performance
(right now it's overfitting).

Also, just as a note-to-self for later, the performance of the main loop
function in `kat_nn.py` can be improved by doing all the p1 predictions
in one call of model.predict() (same for p2 predictions). Then, find the
MAX of the resulting prediction array, and output the character to whom
that corresponds.